Added some boiler plate code for easier parsing of input.
Algorithm is as follows:
Given the number of search engines, we find how many times we fill up a
SET of those distinct engines. Every time we fill a SET with N engines
we will need a switch.  We accomplish this by carrying a tuple of items
during a fold left.