Codejam 2015
I didn't get enough time to devote this year.  I got the first problem,
but then my attempts on the second and fourth revealed some edge case
bugs I didn't account for.  Should be simple enough to go back and fix,
but with kids and wife, not enough time to fix them during the
competition.